requirements.txt
fastapi
uvicorn
transformers>=4.40.0
pillow
python-multipart

numpy<2

torch>=2.6.0
--extra-index-url https://download.pytorch.org/whl/cpu
